% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{dat}
\alias{dat}
\title{A demo dataset}
\format{
A list with 2 items:
\describe{
\item{x}{Covariate matrix}
\item{y}{Outcome vector}
}
}
\usage{
data(dat)
}
\description{
A simulated dataset with covariate matrix \code{x} of size 50 x 10 and an outcome vector \code{y} of length 50.
}
\details{
\code{x} is generated from standard normal distribution. \code{y} is generated from Tweedie distribution with mean equal to exp(sin(x) \%*\% (6, -4, 3, 2, -2,0,...,0)). Only the first five variables are associated with the outcome.
}
\keyword{datasets}
