% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktweedie-package.R
\docType{package}
\name{ktweedie-package}
\alias{ktweedie}
\alias{ktweedie-package}
\title{ktweedie: 'Tweedie' Compound Poisson Model in the Reproducing Kernel Hilbert Space}
\description{
Kernel-based 'Tweedie' compound Poisson gamma model using high-dimensional predictors for the analyses of zero-inflated response variables. The package features built-in estimation, prediction and cross-validation tools and supports choice of different kernel functions. For more details, please see Yi Lian, Archer Yi Yang, Boxiang Wang, Peng Shi & Robert William Platt (2023) \doi{10.1080/00401706.2022.2156615}.
}
\author{
\strong{Maintainer}: Yi Lian \email{yi.lian@mail.mcgill.ca}

Authors:
\itemize{
  \item Archer Yi Yang [copyright holder]
  \item Boxiang Wang
  \item Peng Shi
  \item Robert W. Platt
}

}
\keyword{internal}
