% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktd_estimate.R
\name{ktd_estimate}
\alias{ktd_estimate}
\title{Estimate kernel Tweedie model coefficients}
\usage{
ktd_estimate(
  x,
  y,
  kern,
  lam1,
  rho = 1.5,
  ftol = 1e-08,
  partol = 1e-08,
  abstol = 0,
  maxit = 1e+06,
  sparsity = FALSE,
  lam2 = 0,
  innerpartol = 1e-06,
  innermaxit = 1e+06,
  verbose = FALSE
)
}
\arguments{
\item{x}{Covariate matrix.}

\item{y}{Outcome vector (e.g. insurance cost).}

\item{kern}{Choice of kernel. See \code{\link{dots}} for details on supported kernel functions. It can also simply be a matrix for the {ktweedie}.}

\item{lam1}{A vector of candidate regularization coefficients used in cross-validation.}

\item{rho}{The power parameter of the Tweedie model. Default is 1.5 and can take any real value between 1 and 2.}

\item{ftol}{Stopping criterion based on objective function value. Default is 1e-8.}

\item{partol}{Stopping criterion based on the coefficient values. Default is 1e-8.}

\item{abstol}{Stopping criterion based on absolute value of the objective function. Default is 0.}

\item{maxit}{Maximum number of iterations.}

\item{sparsity}{Logical If true, the \code{sktweedie} model with variable selection will be used. Default is false, for the \code{ktweedie} model.}

\item{lam2}{Regularization coefficient for the sparsity-inducing penalty in the \code{sktweedie} model.}

\item{innerpartol}{Stopping criterion for the inner loops that update kernel parameters and weights based on the coefficient values.}

\item{innermaxit}{Maximum number of iterations for the inner loops that update kernel parameters and variable weights. See Details.}

\item{verbose}{Logical indicating whether to show details of each update.}
}
\description{
\code{ktd_estimate()} estimates the coefficients of the \code{ktweedie} and the \code{sktweedie} model (s for sparse). The latter has an integrated feature selection component that induces sparsity by applying weights on the features and penalizing the weights.
}
\details{
\code{ktd_estimate()} stops when the absolute difference between the objective function values of the last two updates is smaller than \code{ftol}, or the sum of absolute differences between the coefficients of the last two updates is smaller than \code{partol}, before \code{maxit} is reached.
}
\examples{
lam1.seq <- c(1e-5, 1e-4, 1e-3, 1e-2, 1e-1, 1e0, 1e1)
( fit <- ktd_estimate(x = x, y = y,
                      kern = rbfdot(sigma = 1e-8),
                      lam1 = lam1.seq) )

}
