% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktd_predict.R
\name{ktd_predict}
\alias{ktd_predict}
\title{Predict outcome using estimated kernel Tweedie model coefficients}
\usage{
ktd_predict(x.test, model, which.hyperparam = 1, type = "link", KernelMatrix)
}
\arguments{
\item{x}{Covariate matrix.}

\item{fit}{Fitted model from \code{\link{TD_estimate}}}
}
\description{
\code{TD_predict()} predicts the coefficients of the Tweedie compound poisson gamma model in the RKHS. Input is the estimated coefficients, covariate matrix of new data whose outcomes are to be predicted.
}
\details{
\code{TD_predict()} uses the fitted model from \code{\link{TD_estimate}} to estimate the mean claim cost for new data points.
}
\examples{
kern <- rbfdot(sigma = 0.01)
# X: N * p matrix
\dontrun{fit <- TD_estimate(x = X, z = Z, kern = kern,
                            LamReg = 10, phi = 1.8,
                            rho = 1.2, ftol = 1e-10,
                            partol = 1e-15, maxit = 1e6,
                            quiet = TRUE)}
# newdata: n * p matrix
\dontrun{pred <- TD_predict(x = newdata, model = fit)}

}
